+++ example/lib/screens/cart_tab.dart
@@ -5,8 +5,6 @@ import 'package:example/constants.dart';
 import 'package:example/extension.dart';
 import 'package:example/screens/checkout_webview.dart';
 import 'package:flutter/material.dart';
-import 'package:shopify_flutter/mixins/src/shopify_error.dart';
-import 'package:shopify_flutter/models/src/cart/inputs/attribute_input/attribute_input.dart';
 import 'package:shopify_flutter/shopify_flutter.dart';
 
 /// The first variant that can actually be bought, or null if none can.